Move grade type tests and shared test code to seperate files.
